Service Layer V1, Model Constructor
Questions
Should I make a function that returns the sum of upvotes and down votes for the tip?
How do we want data from GETs returned? I can make a json function.
Do we need a get rating function that does the average of all the ratings for a tip in UserRating? Prob not.
Added constructors to models. May not need models for classes that do not have GETs.